refactor(Package/Structure): Redesign package structure
