--- src/main/kotlin/minesweeper/model/board/BoardState.kt
@@ -1,10 +1,0 @@
-package minesweeper.model.board
-
-enum class BoardState {
-    RUNNING,
-    COMPLETED,
-    MINE_EXPLODED;
-
-    val isFinished: Boolean
-        get() = this == COMPLETED || this == MINE_EXPLODED
-}
