+++ src/main/kotlin/minesweeper/model/coordinate/Area.kt
@@ -1,4 +1,4 @@
-package minesweeper.model.board.coordinate
+package minesweeper.model.coordinate
 
 interface Area : Iterable<Position> {
