+++ src/main/kotlin/minesweeper/view/input/ConsoleInputView.kt
@@ -1,7 +1,7 @@
 package minesweeper.view.input
 
 import minesweeper.model.board.Board
-import minesweeper.model.board.coordinate.Position
+import minesweeper.model.coordinate.Position
 import minesweeper.view.input.parser.PositionInputParser
 
 class ConsoleInputView : InputView {
