+++ src/main/kotlin/minesweeper/view/input/InputView.kt
@@ -1,7 +1,7 @@
 package minesweeper.view.input
 
 import minesweeper.model.board.Board
-import minesweeper.model.board.coordinate.Position
+import minesweeper.model.coordinate.Position
 
 fun interface InputView {
