+++ src/test/kotlin/minesweeper/controller/MineSweeperTest.kt
@@ -1,9 +1,9 @@
 package minesweeper.controller
 
 import minesweeper.model.board.Board
-import minesweeper.model.board.Cell
-import minesweeper.model.board.coordinate.BoardArea
-import minesweeper.model.board.coordinate.Position
+import minesweeper.model.cell.Cell
+import minesweeper.model.coordinate.BoardArea
+import minesweeper.model.coordinate.Position
 import minesweeper.view.output.OutputView
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
