+++ src/test/kotlin/minesweeper/model/cell/CellBuilderTest.kt
@@ -1,7 +1,7 @@
-package minesweeper.model.board
+package minesweeper.model.cell
 
 import minesweeper.fixture.toCellBuilder
-import minesweeper.model.board.coordinate.Position
+import minesweeper.model.coordinate.Position
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertAll
